fix: next level exp

+++ MiniGamesBox Classic/src/main/java/plugily/projects/minigamesbox/classic/user/UserManager.java
@@ -111,7 +111,7 @@ public class UserManager {
 
       int level = (int) user.getStatistic(plugin.getStatsStorage().getStatisticType("LEVEL"));
 
-      user.setStatistic(nextLevelExp, (int) Math.ceil(Math.pow(50.0 * level, 1.5)));
+      user.setStatistic(nextLevelExp, (long) Math.ceil(Math.pow(50.0 * level, 1.5)));
 
       //Arena can be null when player has left the arena before this message is retrieved.
       if(arena != null)
